move tests to its own file
